paas-postgresql/api/requirements.txt
flask==2.3.3
kubernetes==28.1.0
PyYAML==6.0.1
requests==2.31.0
